Player added

+++ src/utils/utils.hpp
@@ -2,4 +2,4 @@
 const int SCALE{4};
 
 // Set the size of width and height of the window
-int windowDimension[2]{480, 480};
+float windowDimension[2]{480.f, 480.f};
